# Conflicts:
#	2048_1/Assets/Scripts/FSM.meta
#	2048_1/Assets/Scripts/InputController.cs
#	2048_1/Assets/Scripts/InputController.cs.meta
#	2048_1/Library/ArtifactDB
#	2048_1/Library/Artifacts/07/07255c1b9b9f201a39f3fe8d5ce9f26e
#	2048_1/Library/Artifacts/22/22c4d33a0f5074f90766cb4c2edf8613
#	2048_1/Library/Artifacts/7d/7dfde592e651cf60eca75a07e2045659
#	2048_1/Library/Artifacts/9e/9e271ce0119a143dfede8e35c2ad41bc
#	2048_1/Library/Artifacts/fc/fc14eb0019caea00f1f653a78a6f8b72
#	2048_1/Library/CurrentLayout-default.dwlt
#	2048_1/Library/LastBuild.buildreport
#	2048_1/Library/ScriptAssemblies/Assembly-CSharp.dll
#	2048_1/Library/SourceAssetDB
#	2048_1/Library/com.unity.addressables/AddressablesBuildTEP.json
#	2048_1/obj/Debug/Assembly-CSharp-firstpass.csproj.AssemblyReference.cache
#	2048_1/obj/Debug/Assembly-CSharp.csproj.AssemblyReference.cache
#	2048_1/obj/Debug/Unity.2D.Animation.Samples.Dependency.csproj.AssemblyReference.cache
#	2048_1/obj/Debug/Unity.2D.Animation.Samples.Editor.csproj.AssemblyReference.cache
#	2048_1/obj/Debug/Unity.2D.Animation.Samples.Runtime.csproj.AssemblyReference.cache

+++ 2048_1/Assets/Scripts/WindowManager.cs
@@ -8,7 +8,7 @@ public class WindowManager : MonoBehaviour
     public void Show()
     {
         this.gameObject.SetActive(true);
-        this.gameObject.transform.DOShakeScale((float)1, (float)0.5, 5);
+        this.gameObject.transform.DOShakeScale(1f, 0.5f, 5);
     }
 
     public void Hide()
